Renames project to ExtensibleDataPipeline (#10)
* rename ExtensibleMessageBroker to ExtensibleDataPipeline

* Renames solution to Edp

Refactors the solution, projects, and namespaces from 'Emb' to 'Edp' to align with the new project naming convention.

This change ensures consistency across the codebase and improves maintainability.

fix

* Renames solution in README files.

Updates README files to reflect the solution's rename, ensuring accurate naming and branding across the documentation.

This change aligns the README with the new project name.

* Updates project references and names

Renames project-related variables and references in the release workflow file to reflect the renaming of the solution from ExtensibleMessageBroker to ExtensibleDataPipeline.

This ensures that the CI/CD pipeline uses the correct project paths and file prefixes for building and releasing the application. It also updates the checkout and upload-artifact actions to newer versions.

* enable NRT, fix xmlns, dispose objects

* fix warnings